Incorporate the stateless users.conf file for LightDM
Clear Linux now uses a $(datadir)/lightdm/users.conf file in the absence of
the admin-provided /etc/lightdm/users.conf, in line with the stateless
implementation of the system.

In using the users.conf implementation for the user-list element of LightDM,
we can avoid an unnecessary dependency on AccountsService simply for list
filtering.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>