schemas: Introduce GSettings overrides for glib/dconf defaults
In order to attain easy maintenance and control over the default appearance,
behaviour, and function, of the GNOME components, we add gschema overrides
which will control all of these aspects.

When the glib-compile-schemas unit is run, these overrides become the
default GSettings, without requiring any patches in the upstream packages.

Due to the file lengths now incorporated, we must switch to a more modern
format of tarball to ensure distcheck continues to function as expected.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>