schemas: Enable GTK animations by default once more
This will enable animations in GTK and GNOME Shell again. This change has
to now live as a gschema override as our default schema was patched around
1 year ago to forcibly disable animations, leading to a degraded user
experience on local installations.

The plan will be to revert the patch, and introduce the gschema override to
then disable animations at a higher priority override level for VNC session
bundles, so that we don't attempt to follow "one size fits all" for these
different scenarios.

Signed-off-by: Ikey Doherty <ikey.doherty@intel.com>